moved authenticateUserModal to gallery
